ledger hit

+++ database/migrations/2025_12_15_124920_create_sale_return_data_table.php
@@ -14,17 +14,17 @@ return new class extends Migration
         Schema::create('sale_return_data', function (Blueprint $table) {
             $table->id();
             $table->foreignId("sale_return_id")->constrained("sales_return")->cascadeOnDelete();
-            $table->unsignedInteger("quantity");
-            $table->unsignedInteger("rate");
-            $table->unsignedInteger("gross_amount");
-            $table->unsignedInteger("discount_percent");
-            $table->unsignedInteger("discount_amount");
-            $table->unsignedInteger("amount");
-            $table->unsignedInteger("gst_percentage");
-            $table->unsignedInteger("gst_amount");
-            $table->unsignedInteger("net_amount");
-            $table->unsignedInteger("line_desc");
-            $table->unsignedInteger("truck_no");
+            $table->float("quantity");
+            $table->float("rate");
+            $table->float("gross_amount");
+            $table->float("discount_percent");
+            $table->float("discount_amount");
+            $table->float("amount");
+            $table->float("gst_percentage");
+            $table->float("gst_amount");
+            $table->float("net_amount");
+            $table->float("line_desc");
+            $table->float("truck_no");
             $table->string("packing");
             $table->unsignedInteger("no_of_bags");
             $table->foreignId("sale_invoice_data_id")->constrained("sales_invoice_data")->cascadeOnDelete();
